feat: trigger summary block on any video, remove content pointer
We will be building a path for ai-spot to get content from edx rather
than having it pushed by the browser, including video transcripts.

To demo that quickly we're going to populate selected course content
into a temp store in ai-spot and it will fetch from a local fake edx.

So this change removes the local content indicator and triggers on any
video and summaries will for a while only work on the pre-fetched
while we catch up the rest of the content path.